test(contrib): consolidate new baggage security tests into existing files
Per review, fold the tests added for the ot-baggage-* propagation
hardening into the packages' existing test files instead of one-off
files: contrib/google.golang.org/grpc/grpc_test.go and
contrib/net/http/roundtripper_test.go. No behavioral change.
